requirements.txt
discord.py==2.5.2
yt_dlp==2025.2.19
PyNaCl==1.5.0